@sugyan 2022-03-07
benchmark:
MacBook Pro (13-inch, 2019, Four Thunderbolt 3 ports)
2.8 GHz Quad-Core Intel Core i7
16 GB 2133 MHz LPDDR3

before:

```
test tests::bench_legal_moves_from_default ... bench:         878 ns/iter (+/- 115)
test tests::bench_legal_moves_maximum      ... bench:       4,171 ns/iter (+/- 495)
test tests::bench_perft_from_default       ... bench: 597,462,520 ns/iter (+/- 50,367,964)
test tests::bench_perft_from_maximum_moves ... bench: 384,189,151 ns/iter (+/- 17,422,152)
```

after:
```
test tests::bench_legal_moves_from_default ... bench:         879 ns/iter (+/- 69)
test tests::bench_legal_moves_maximum      ... bench:       4,190 ns/iter (+/- 487)
test tests::bench_perft_from_default       ... bench: 546,657,227 ns/iter (+/- 40,266,998)
test tests::bench_perft_from_maximum_moves ... bench: 374,914,346 ns/iter (+/- 17,988,295)
```